services/py-api/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Ivan Obreshkov <ivanobreshkov12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
fastapi = "^0.114.2"
uvicorn = "^0.30.6"
httpx = "^0.27.2"
structlog = "^24.4.0"
result = "^0.17.0"
pymongo = "^4.8.0"
motor = "^3.6.0"
resend = "^2.4.0"
pydantic = { extras = ["email"], version = "^2.9.2" }
pyjwt = "^2.10.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
locust = "^2.31.8"

[tool.poetry.scripts]
start = "src.server.app_entrypoint:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
